# device/api.yaml
openapi: 3.0.3
info:
  title: Smart Home Device API
  description: API to manage and retrieve smart home devices.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /devices:
    get:
      summary: List all devices
      operationId: getDevices
      description: Retrieve a list of all registered smart home devices.
      responses:
        "200":
          description: A list of smart home devices.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
        "500":
          description: Server error

  /device/{id}:
    get:
      summary: Get device
      operationId: GetDevice
      description: Retrieve a single device by its ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A single device.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Device not found
        "500":
          description: Server error

  /device/{id}/configurations:
    get:
      summary: Get device configurations
      operationId: GetDeviceConfigurations
      description: Retrieve a list of configurations for a device.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A list of device configurations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConfigurationListItem"
        "404":
          description: Device not found
        "500":
          description: Server error
    post:
      summary: Create device configuration
      operationId: CreateConfiguration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        description: Device configuration to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Configuration"
        required: true
      responses:
        "201":
          description: Configuration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Configuration"
        "400":
          description: Invalid configuration
        "404":
          description: Device not found

  /device/{id}/configuration/{configurationId}:
    get:
      summary: Get device configuration
      operationId: GetConfiguration
      description: Retrieve a single configuration for a device.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: configurationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A single configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Configuration"
        "404":
          description: Configuration or device not found
        "500":
          description: Server error
    put:
      summary: Update device configuration
      operationId: UpdateConfiguration
      description: Update a single configuration for a device.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: configurationId
          required: true
          schema:
            type: string
      requestBody:
        description: Configuration to update
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Configuration"
        required: true
      responses:
        "200":
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Configuration"
        "400":
          description: Invalid configuration
        "404":
          description: Device not found
        "500":
          description: Server error

  /device/{id}/configuration/{configurationId}/status:
    post:
      summary: Toggle device configuration status
      operationId: ToggleConfigurationStatus
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: configurationId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Configuration status toggled
        "404":
          description: Configuration or device not found
        "500":
          description: Server error

components:
  schemas:
    Device:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the device
        name:
          type: string
          description: Human-readable name of the device
        type:
          type: string
          description: Type of the smart device (e.g., light, thermostat)
        status:
          type: string
          description: Current status of the device (e.g., on, off, idle)
      required:
        - id
        - name
        - type
        - status

    ConfigurationListItem:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the configuration
        name:
          type: string
          description: Human-readable name of the configuration
        active:
          type: boolean
          description: Whether the configuration is active or not
      required:
        - id
        - name
        - active

    Configuration:
      type: object
      allOf:
        - $ref: "#/components/schemas/ConfigurationListItem"
        - type: object
          properties:
            data:
              type: object
              description: Configuration data